feat: separate fetcher from por main file
